src/conftest.py
import pytest
from django.contrib.auth import get_user_model
from model_bakery import baker
from rest_framework.test import APIClient

User = get_user_model()


@pytest.fixture
def api_client():
    return APIClient


@pytest.fixture
def user():
    """default user"""
    return baker.make(User)
